requirements.txt
# For GPU: pip uses this index for CUDA builds of PyTorch
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio
pyyaml>=6.0
pypdf>=4.1.0
python-docx>=1.1.0
rank-bm25>=0.2.1
numpy
langchain-core>=0.2.0,<0.3
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface>=0.0.3
langchain-text-splitters==0.2.4
beautifulsoup4>=4.12.0
sentence-transformers
transformers
faiss-cpu
huggingface-hub
langsmith
python-dotenv
fpdf2
huggingface-hub